pyproject.toml
[tool.poetry]
name = "aura"
version = "1.17.1"
description = "Karma Discord Bot"
authors = ["Georg Göring <g.goering@protonmail.com>"]
[tool.poetry.dependencies]
python = "^3.8.6"
pymongo = "^3.11.0"
PyYAML = "^5.3.1"
"discord.py" = "^1.5.0"
[tool.poetry.dev-dependencies]
pytest = "^6.1.1"
mongomock = "^3.20.0"
flake8 = "^3.8.4"
[tool.poetry.scripts]
aura = 'aura.bot:main'